+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Add check that every property provides one or more examples with the `examples` keyword.
+- Add check that every deprecated property has a comment.
+- Add check that properties don't provide more than 5 examples.
+
+## [0.7.0] - 2023-02-06
+
 - Add check that every array specifies the schema of its items with the `items` keyword.
 - Simplify output of `schemalint verify` to improve appearance in GitHub actions log.
 - Add check that `additionalProperties` is disabled on all objects.
@@ -53,7 +58,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added first basic linting.
 
-[Unreleased]: https://github.com/giantswarm/schemalint/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/giantswarm/schemalint/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/giantswarm/schemalint/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/giantswarm/schemalint/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/giantswarm/schemalint/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/giantswarm/schemalint/compare/v0.3.0...v0.4.0